CVS_SILENT
Corrected 'comparison between signed and unsigned' warning (using a cast).

svn path=/trunk/kdetoys/ktux/; revision=319032

+++ spriteanim.cpp
@@ -132,7 +132,7 @@ void SpriteGroup::next()
 //---------------------------------------------------------------------------
 void SpriteGroup::refresh()
 {
-    if (mActive.count() < mCount)
+    if (((int) mActive.count()) < mCount)
     {
         SpriteObject *sprite = mAvailable.first()->create(mCanvas);
         mActive.append(sprite);
